refactor: 인터페이스 제거, 직접 클래스를 활용하게 변경
