+++ src/main/java/com/example/msglab/domain/MessageClientResp.java
@@ -1,6 +1,24 @@
 package com.example.msglab.domain;
 
-public interface MessageClientResp {
-    public Status getStatus();
-    public String getId();
+import org.springframework.http.ResponseEntity;
+
+/**
+ * FCM 등으로 push message를 전송하면 리턴되는 응답 객체입니다.
+ */
+public class MessageClientResp {
+
+    private final ResponseEntity<String> response;
+
+    public MessageClientResp(ResponseEntity<String> response) {
+        this.response = response;
+    }
+    // todo(hun) : reponse에 담긴 httpCode를 Status로 변환하기
+    public Status getStatus() {
+        return Status.OK;
+    }
+
+    // todo(hun) : response body에 담긴 json 형태의 아이디 : 값 매핑을 String타입의 아이디로 변환하기
+    public String getId() {
+        return response.getBody();
+    }
 }
